// frontend/src/app/core/api/model/exportResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExportResponse { 
    fileName?: string;
    format?: string;
    entryCount?: number;
    encrypted?: boolean;
    data?: string;
    exportedAt?: string;
}